requirements.txt
streamlit==1.33.0
#mysql-connector-python== 8.3.0
#bcrypt==4.1.3
#scikit-learn==1.2.2
langchain==0.2.3
langchain-openai==0.1.8
langchain-community==0.2.3
PyMuPDF==1.24.7
langdetect==1.0.9
langchain_core==0.2.22
chromadb==0.5.3
python-dotenv==1.0.1